[design] about페이지로 이동(text페이지 삭제)
